[TASK] Move Ajax controllers to Controller namespace
Fixes #156
